+++ include/boost/unordered/detail/foa/core.hpp
@@ -29,7 +29,7 @@
 #include <boost/unordered/detail/mulx.hpp>
 #include <boost/unordered/detail/static_assert.hpp>
 #include <boost/unordered/detail/type_traits.hpp>
-#include <boost/unordered/unordered_printers.hpp>
+#include <boost/unordered/detail/unordered_printers.hpp>
 #include <climits>
 #include <cmath>
 #include <cstddef>
